Delete patient_join_result.dart improve the class BackendPatientListApi
